pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-web-search"
version = "0.1.0"
description = "MCP server providing curated web search tooling."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27.0",
    "anyio>=4.3.0",
    "tenacity>=8.2.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.7.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.1.0",
    "prometheus-client>=0.20.0",
    "opentelemetry-sdk>=1.26.0",
    "opentelemetry-exporter-otlp>=1.26.0",
    "beautifulsoup4>=4.12.0",
    "mcp>=1.18.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.5.5",
    "mypy>=1.9.0",
    "uvicorn[standard]>=0.30.0",
    "fastapi>=0.111.0"
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.10"
strict = true
